Faz a separação de responsabilidades.
