--- gastos-query-api/src/main/java/br/com/santander/gastos/integracao/IntegracaoApplication.java
@@ -1,16 +1,0 @@
-package br.com.santander.gastos.integracao;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class IntegracaoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(IntegracaoApplication.class, args);
-    }
-
-}
-
