--- gastos-query-api/src/main/java/br/com/santander/gastos/integracao/dto/GastoAdicionadoResponse.java
@@ -1,7 +1,0 @@
-package br.com.santander.gastos.integracao.dto;
-
-import java.io.Serializable;
-
-public class GastoAdicionadoResponse implements Serializable {
-
-}
