--- gastos-query-api/src/main/java/br/com/santander/gastos/integracao/exceptions/GastoNaoEncontradoException.java
@@ -1,4 +1,0 @@
-package br.com.santander.gastos.integracao.exceptions;
-
-public class GastoNaoEncontradoException extends RuntimeException{
-}
