+++ gastos-query-api/src/main/java/br/com/santander/gastos/query/repository/GastoRepository.java
@@ -1,6 +1,6 @@
-package br.com.santander.gastos.integracao.repository;
+package br.com.santander.gastos.query.repository;
 
-import br.com.santander.gastos.integracao.entity.GastoEntity;
+import br.com.santander.gastos.query.entity.GastoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
