+++ gastos-query-api/src/test/java/br/com/santander/gastos/query/IntegracaoApplicationTests.java
@@ -1,4 +1,4 @@
-package br.com.santander.gastos.integracao;
+package br.com.santander.gastos.query;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
